+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 * It is now possible to `join` BioSymbols into a BioSequence.
+* Added non-exported functions `trydecode` and `tryencode`.
 
 ## [3.0.1]
 ### Removed